+++ src/components/Slider/Slider.tsx
@@ -12,7 +12,6 @@ import {
   SliderFilledTrack as ChakraSliderFilledTrack,
   SliderThumb as ChakraSliderThumb,
   SliderTrack as ChakraSliderTrack,
-  useColorMode,
   useMultiStyleConfig,
 } from "@chakra-ui/react";
 import React, { forwardRef, InputHTMLAttributes } from "react";
@@ -137,7 +136,6 @@ export const Slider: ChakraComponent<
     }
 
     const [currentValue, setCurrentValue] = useStateWithDependencies(value);
-    const isDarkMode = useColorMode().colorMode === "dark";
 
     let finalIsInvalid = isInvalid;
     // In the Range Slider, if the first value is bigger than the second value,
@@ -155,7 +153,6 @@ export const Slider: ChakraComponent<
       );
     }
     const styles = useMultiStyleConfig("ReservoirSlider", {
-      isDarkMode,
       isDisabled,
       isInvalid: finalIsInvalid,
       showBoxes,