+++ src/components/Tabs/Tabs.tsx
@@ -76,8 +76,8 @@ const getElementsFromData = (
     const tempTab = (
       <Tab
         fontSize={{
-          base: "mobile.subtitle.subtitle1",
-          md: "desktop.subtitle.subtitle1",
+          base: "mobile.tab.tab1",
+          md: "desktop.tab.tab1",
         }}
         key={index}
         onClick={useHash ? () => onClickHash(`tab${index + 1}`) : undefined}